requirements.txt
# Optional: HHDG frame ↔ gallery color matching (enrich_product)
Pillow>=10.0.0
